Restructuring the application
