+++ recipeapp/src/Recipes/Root/Root.css
@@ -1,0 +1,3 @@
+h2{
+    margin: 5% 40%;
+}
