+++ recipeapp/src/Recipes/Root/index.js
@@ -1,0 +1,2 @@
+export { Root } from './Root';
+
