+++ recipeapp/src/Recipes/SearchBar/SearchBar.js
@@ -1,26 +1,21 @@
 import React,{Component} from 'react';
-import '../SearchBar/SearchBar.css'
+import './SearchBar.css'
 
-class SearchBar extends Component {
+export class SearchBar extends Component {
     state = {
         searchValue: '',
     }
-
     handleChange = (event) => {
         this.setState({
             searchValue: event.target.value
         })
-
     }
-
     render(){
         return(
             <div className="searchBar">
                 <input type="text" className = "searchInput" onChange={this.handleChange}/>
             </div>
-
         );
     }
 }
 
-export default SearchBar;
