+++ recipeapp/src/Recipes/SearchBar/index.js
@@ -1,0 +1,1 @@
+export { SearchBar } from './SearchBar';
