+++ recipeapp/src/Recipes/index.js
@@ -1,17 +1,1 @@
-import React, {Component} from 'react';
-import SearchBar from './SearchBar/SearchBar';
-import '../Recipes/recipes.css'
-
-class Recipes extends Component {
-    render(){
-        return(
-            <div>
-                <h2>Search Recipes</h2>
-                <SearchBar/>
-            </div>
-
-        );
-    }
-}
-
-export default Recipes
+export { Root as Recipes } from './Root';
