--- recipeapp/src/Recipes/recipes.css
@@ -1,3 +1,0 @@
-h2{
-    margin-top: 5%;
-}
